+++ tests/test_pysftp_02.py
@@ -26,6 +26,29 @@ SFTP_LOCAL = {'host':'localhost', 'username':'test', 'password':'test1357'}
 skip_if_ci = pytest.mark.skipif(os.getenv('CI', '')>'', reason='Not Local')
 
 
+@skip_if_ci
+def test_chmod_not_exist():
+    '''verify error if trying to chmod something that isn't there'''
+    with pytest.raises(IOError):
+        with pysftp.Connection(**SFTP_LOCAL) as sftp:
+            sftp.chmod('i-do-not-exist.txt', 666)
+
+@skip_if_ci
+def test_chmod_simple():
+    '''test basic chmod'''
+    new_mode = 711
+    with tempfile_containing('') as fname:
+        base_fname = os.path.split(fname)[1]
+        with pysftp.Connection(**SFTP_LOCAL) as sftp:
+            org_attrs = sftp.put(fname)
+            sftp.chmod(base_fname, new_mode)
+            new_attrs = sftp.stat(base_fname)
+            sftp.remove(base_fname)
+    # that the new mod 711 is as we wanted
+    assert pysftp.st_mode_to_int(new_attrs.st_mode) == new_mode
+    # that we actually changed something
+    assert new_attrs.st_mode != org_attrs.st_mode
+
 def test_path_retreat():
     '''test path_retreat generator'''
     pth = 'foo/bar/baz'