+++ docs/changes.rst
@@ -4,8 +4,10 @@ Change Log
 
 * 0.2.6 (dev)
 
-  * added preserve_mtime parameter to .put() method, optionally updates the remote file's st_mtime to match the local file.
+  * added preserve_mtime parameter to .put() function, optionally updates the remote file's st_mtime to match the local file.
   * added preserve_mtime parameter to .get() method, optionally updates the local file's st_mtime to match the remote file
+  * added .exists() and .lexists() function, use .stat() and .lstat() respectively
+  * added .symlink() method
 
 * 0.2.5 (current, released 2014-05-15)