+++ docs/changes.rst
@@ -9,6 +9,7 @@ Change Log
   * added ``.exists()`` and ``.lexists()`` function, use .``.stat()`` and ``.lstat()`` respectively
   * added ``.symlink()`` method
   * added ``.isdir()``, ``.isfile()``, ``.makedirs()``
+  * added ``.chmod()``
 
 * 0.2.5 (current, released 2014-05-15)