+++ docs/changes.rst
@@ -10,6 +10,7 @@ Change Log
   * added ``.symlink()`` method
   * added ``.isdir()``, ``.isfile()``, ``.makedirs()``
   * added ``.chmod()``
+  * added ``.chown()``
 
 * 0.2.5 (current, released 2014-05-15)